refactor: Product 도메인의 Repository 정리
