+++ api/src/main/java/com/albert/commerce/product/infra/persistence/ProductDaoImpl.java
@@ -1,0 +1,69 @@
+package com.albert.commerce.product.infra.persistence;
+
+import com.albert.commerce.product.command.domain.Product;
+import com.albert.commerce.product.command.domain.QProduct;
+import com.albert.commerce.product.query.ProductDao;
+import com.albert.commerce.store.command.domain.QStore;
+import com.albert.commerce.store.command.domain.StoreId;
+import com.albert.commerce.user.command.domain.QUser;
+import com.albert.commerce.user.command.domain.UserId;
+import com.querydsl.jpa.JPAExpressions;
+import com.querydsl.jpa.JPQLQuery;
+import com.querydsl.jpa.impl.JPAQuery;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.support.PageableExecutionUtils;
+import org.springframework.stereotype.Repository;
+
+@RequiredArgsConstructor
+@Repository
+public class ProductDaoImpl implements ProductDao {
+
+    private final JPAQueryFactory jpaQueryFactory;
+
+    private static JPQLQuery<UserId> getUserIdByEmail(String userEmail) {
+        QUser qUser = QUser.user;
+        return JPAExpressions
+                .select(qUser.id)
+                .from(qUser)
+                .where(qUser.email.eq(userEmail));
+    }
+
+    private static JPQLQuery<StoreId> getStoreIdByUserId(JPQLQuery<UserId> userIdByEmail) {
+        QStore qStore = QStore.store;
+        return JPAExpressions
+                .select(qStore.storeId)
+                .from(qStore)
+                .where(qStore.storeUserId.userId.eq(
+                        userIdByEmail
+                ));
+    }
+
+    @Override
+    public Page<Product> findProductsByUserEmail(String userEmail, Pageable pageable) {
+
+        QProduct qProduct = QProduct.product;
+        JPQLQuery<StoreId> storeIdQuery = getStoreIdByUserId(getUserIdByEmail(userEmail));
+        List<Product> contentProducts = jpaQueryFactory.selectFrom(qProduct)
+                .where(
+                        qProduct.storeId.eq(
+                                storeIdQuery
+                        )
+                )
+                .offset(pageable.getOffset())
+                .limit(pageable.getPageSize())
+                .fetch();
+        JPAQuery<Long> productJPAQuery = jpaQueryFactory
+                .select(qProduct.count())
+                .from(qProduct)
+                .where(
+                        qProduct.storeId.eq(
+                                storeIdQuery
+                        )
+                );
+        return PageableExecutionUtils.getPage(contentProducts, pageable, productJPAQuery::fetchOne);
+    }
+}
